requirements.txt
# Flask核心组件
Flask==2.3.3
Flask-PyMongo==2.3.0
Flask-CORS==4.0.0
Flask-Caching==2.1.0
Flask-SocketIO==5.3.2
python-socketio==5.7.2
python-engineio==4.3.4

# 数据处理
numpy==1.23.5
pandas==1.5.3
matplotlib==3.7.1
seaborn==0.12.2

# HTTP客户端和HTML解析
requests==2.28.2
beautifulsoup4==4.11.2
lxml==4.9.2

# 环境变量和配置
python-dotenv==1.0.0
PyJWT==2.8.0

# 辅助工具
python-dateutil==2.8.2
pytz==2022.7.1
Werkzeug==2.3.7
itsdangerous==2.1.2
six==1.16.0
pymongo==4.5.0
blinker>=1.6.2

# 其他实用库
Jinja2==3.1.2
MarkupSafe==2.1.2
click==8.1.3
colorama==0.4.6

# 认证与安全
bcrypt==4.0.1

# 测试工具
pytest==7.4.2

# WebSocket
eventlet==0.33.3
gevent==22.10.2

# 其他工具
Pillow==9.4.0   # 图像处理库，用于头像和附件处理
openpyxl==3.1.1 # Excel支持，用于数据导出

# 其他工具
gunicorn==21.2.0
coverage==7.3.1
black==23.7.0
flake8==6.1.0 